Give the session more time each time a restricted page is accessed.
Previously, a session would timeout one hour after the user logged
in.  This could be in the middle of an editing session, and work
would be lost.  Now the session is reset when the user hits the
edit page, so there will be no problem unless the user spends more
than one hour editing a single page.