Added Gitit.ContentTransformer module; removed Gitit.Convert.
Thanks to Anton van Straaten for the patch. The ContentTransformer
module replaces Gitit.Convert. It defines a number of single-purpose
combinators that can be combined to yield various kinds of content
conversions. These are used to define showPage, preview,
showHighlightedSource, and other handlers that used to be
defined in Gitit.hs.