Change preview URL from /_preview to *___preview.
Thanks to Henry Laxen for the idea.  This solves a residual
problem running gitit with mod_proxy_html on an alternative
docroot.  mod_proxy_html doesn't rewrite links in javascripts,
so we changed preview.js to link to "___preview" instead of
"/_preview", and we test for this suffix in the routing code.