Fixed problem with revert (resolves Issue #45).
The changes are in editPage, which was setting the sha1 parameter
on the form to be the same as the revision parameter of the request.
The result was that updatePage took the user to be modifying an
old version of the page, and automatically merged the changes
with the latest version.  So nothing happened.

Solution:  when pRevision is set, it means we're doing a revert,
so we set the sha1 = the revision ID of the latest version of the
page, but get content from the specified revision.