Reintroduced AppState parameter to PageTransform plugins.
However, instead of making the user set it, we have mkPageTransform and
mkPageTransformM handle the plumbing. So users can pretty much ignore
it.

The reason we need to set this is that the state IORef in the plugins
points to an uninitialized state (since the state is only initialized
after gitit starts running). Maybe there is a way around this.

Note that each plugin has, essentially, its own read-only copy of the
state.