Large architecture change.
Instead of using 'handle' and handlers in the
Web monad, we now use happstack's routing combinators
and handlers in the ServerPart monad.

This should make gitit easier to extend, and is a
necessary step on the way to gitit as a library.

Added getPage instead of passing page as an argument
to handlers.  For now, we still pass params, but this
may change.

Removed unused and out of date Makefile.