Redirect /wiki to /wiki/ when gitit served at /wiki.
Reason:  if the base url is /wiki, and you are browsing /wiki,
then wikilinks that are relative URLS will not work correctly.
For example, "Help" will take you to /Help rather than
/wiki/Help.  The simple solution is to ensure that users
are always at /wiki/ rather than /wiki.

This has no effect when gitit is being served at the root url.

Thanks to MightyByte for pointing out the problem.