Store renderPage function in gitit state.
This is more flexible than storing a page template, since the
user may want to use a custom page rendering function, even
one not based on string templates.

Removed old Template data structure.  We still store the path
of the templates directory, as it is needed by the reloadTemplates
action.