Tweaked build-depends to help 'cabal install'.
cabal install thinks it has to install an older version of network unless the
network lower bound is raised.  So we raise it for ghc 7.0.3, so that
the version of network packaged with it in the Haskell platform is used.
This is a kludge; cabal install should be smarter, but we can work
around it.