Add the new configuration option 'absolute-urls'.
When turned on, this makes wikilinks absolute w.r.t. the base-url.
By default, they are relative.

So, for example, in a wiki served at the path 'wiki', on a page
Sub/Page, the wikilink

[Cactus]()

will produce a link to '/wiki/Cactus' if absolute-urls is on,
and otherwise the relative link 'Cactus'.

Patch due to lemmih.