Added user list page at _users
This displays a table of all the users of the site, including the times they last logged in.

By default, the table does not show user's email addresses. I added a config option that enables this which is off by default. When displaying email addresses, it uses pandoc's reference obfuscation.