Revert "Create page if no match is found when using _go"
This reverts commit 1328545e1b9e86132b4937ac5fb2afc314bb0758.

I hadn't fully appreciated the loss of the old prefix-match behavior,
which was useful.  So for now I want to pull this change back.
(See #397 and discussion on gitit-discuss.)

Note that the old behavior copied wikipedia.

The standard way to create pages is to first create a link
to the page in an existing page, then click on the link.