lookahead instead of \b when patt starts with \W
These changes capture both:

- regular patterns that don't start with non-word characters
- patterns that start with non-word characters, even if they appear at start of line

One unfortunate artifact of the changes is that in patterns with spaces, the spaces will be highlighted. But this is the only way (under the current approach) to deal with patterns that start and end with punctuation but also contain spaces. In that case, because the lookahead boundary is enabled, the space needs to remain in the pattern so that the lookahead test works.