Added MTable plugin.
This plugin adds a new table syntax that I find less irritating than
having to hand-align everything.  Example:

``` {.mtable}
| some | table |
| that doesn't have the columns | aligned with each other |
```

is equivalent to:

--------------------------------  ------------------------
some                              table
that does have the columns        aligned with each other
--------------------------------  ------------------------

This is similar to Pandoc's pipe_tables extension but not
all versions of pandoc have that (particularly the one in
Debian 7 stable), so this hack might be useful to someone.