PDF export: fix vulnerability.
Augustin Laville pointed out that it is possible to
leak content from files on the file system in PDF export.
There are two ways of doing this.  One is by simply
including

    \include{/path/to/file}

in the markdown; the raw tex is passed through to the
LaTeX intermediary when producing a PDF.  In many cases
this will lead to a LaTeX syntax error, but there are ways of
injecting TeX includes into a verbatim environment.

The solution is to cause latex to use a local texmf.cnf
file when exporting PDF. The texmf.cnf file can specify
openout_any = p and openin_any = p, which prevents access
to hidden files and files outside the current tree.