Allow using any reader for PageType
Instead of a closed list, allow anything Pandoc understands.  If there is no
markup/markupHelp file for the reader, then there will be no help, otherwise
this works as before.

Instead of special-casing +lhs this is now just a normal extension specification
understood by Pandoc.  +emoji is always added as before. Reader specs are run
through toLower for compatibility with the existing system.

Closes #636
Closes #580