Adjust for hoauth2 >= 2.3.0
hoauth2 is used by Stackage Nightly 2.3.0 and will thus be included in
Stackage LTS 19 when it comes out. Another motivation for this being
able to build with aeson 2.0 which is possible since hoauth2 2.1.

I've elected not to try to support a wider range (say >= 2.0 && < 2.4 or
even >= 1.3.0 && < 2.4) because *every version* released between 2.0 and
2.3 contains a breaking change that affects gitit, the CPP would turn
out to be quite the mess. Since there haven't been any functional
changes since the last release on master, people that can't upgrade yet
can safely skip a (hypothetical) immediate release.

If wanted, support for hoauth2 >= 2.1 && < 2.4 could possibly be
interesting, only supporting aeson >= 2.0 is probably a safe bet for
future releases.