Move everything from xhtml to blaze & HTML5
This is a rather mechanical translation from the old `xhtml` library
to the more modern `blaze` library.

We use HTML5 everywhere, thus we also change the `DOCTYPE`.

Pandoc needs to be configured to return HTML5 as well, but it looks
like every use of the html writer already uses `writeHtml5String`. So
I guess we are good?