feat(Config): Add json config format option
In order to make it easier to generate gitit configuration, we add the
ability for config files to be json values in addition to the custom
config format parser.

This should be 100% backward-compatible, with the error messages being
reasonably clear about the fallback mechanism.

Json strings, numbers and booleans are supported as-is, and the
default section also works (sections are distinguished by their name
being in `[brackets]`).

Potentially a bit more documentation might be needed.